pull tests up to RepositoryTest
